+++ packages/sz-rust-frontend-codegen/src/generators/react.rs
@@ -1,5 +1,4 @@
 //! React 组件生成器
-#![allow(dead_code)]
 
 use crate::config::GenerationConfig;
 use crate::error::FrontendCodegenError;